Refractor chathub
